third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-002-expected.txt


FAIL CSS Test: @viewport constrained - max-width 10000px assert_equals: expected 320 but got 980
